requirements.txt
fastapi[standard]
ingredient_parser_nlp
azure-functions
uvicorn